Skip discarded ACT decoder layers and add runtime kernel choices
DETRVAE.forward consumes only the first stacked decoder output
(`self.transformer(...)[0]` with return_intermediate_dec=True), so decoder
layers past the first never influence predictions and receive exactly-zero
gradients from autograd. `decoder_layers_used = 1` (a plain attribute, not
architecture or state) stops computing them; the trainer feeds those
parameters the same zero gradients so AdamW's decoupled weight decay is
unchanged. Predictions and every gradient are bitwise identical.

nn.MultiheadAttention call sites select need_weights per the layer's
`attention` mode: the explicit upstream matmul+softmax path in fp32 (TF32
tensor-core GEMMs on Blackwell beat the fp32-only fused kernel) and fused
scaled-dot-product kernels under autocast. FrozenBatchNorm2d applies its
affine in the activation dtype (no-op in fp32) so bf16 autocast activations
are not promoted; the latent distribution and output heads stay fp32 under
autocast. The padding mask is created on the target device and AdamW can
use the fused CUDA implementation (`fused_optimizer` runtime override).

+++ detr/models/backbone.py
@@ -54,7 +54,9 @@ class FrozenBatchNorm2d(torch.nn.Module):
         eps = 1e-5
         scale = w * (rv + eps).rsqrt()
         bias = b - rm * scale
-        return x * scale + bias
+        # Keep the affine in the activation dtype (no-op in fp32) so autocast bf16 activations are not
+        # promoted back to fp32 here; the frozen statistics stay in fp32.
+        return x * scale.to(x.dtype) + bias.to(x.dtype)
 
 
 class BackboneBase(nn.Module):
